update models swap int to bigint in pk, created like repository, optimized queries repositories

+++ app/database/models/comment.py
@@ -1,4 +1,4 @@
-from sqlalchemy import String, Integer, ForeignKey, TIMESTAMP, func
+from sqlalchemy import String, BIGINT, ForeignKey, TIMESTAMP, func
 from sqlalchemy.orm import relationship, Mapped, mapped_column
 from datetime import datetime
 
@@ -8,9 +8,9 @@ from .base import Base
 class Comment(Base):
     __tablename__ = 'comments'
 
-    id: Mapped[int] = mapped_column(Integer, primary_key=True)
-    user_id: Mapped[int] = mapped_column(Integer, ForeignKey('users.id'), index=True)
-    post_id: Mapped[int] = mapped_column(Integer, ForeignKey('posts.id'), index=True)
+    id: Mapped[int] = mapped_column(BIGINT, primary_key=True)
+    user_id: Mapped[int] = mapped_column(BIGINT, ForeignKey('users.id'), index=True)
+    post_id: Mapped[int] = mapped_column(BIGINT, ForeignKey('posts.id'), index=True)
     text_content: Mapped[str] = mapped_column(String(1024), nullable=False)
     created_at: Mapped[datetime] = mapped_column(TIMESTAMP, server_default=func.now(), nullable=False, index=True)
